Add metric info for: multiplexing and period
Multiplexed event can be inaccurate, especially if it's combined with
frequency-based sampling. It is important to notify users that the
metric is based on multiplexed event or not. If it's multiplexed (or
it's frequency-based sampling), we should inform the average of sampling
period and its standard deviation.

Minor update: reorganize the code for customized event (such as kernel
blocking)

+++ src/tool/hpcrun/sample-sources/perf/perfmon-util.c
@@ -68,6 +68,7 @@
  *****************************************************************************/
 
 #include <hpcrun/messages/messages.h>
+#include "perf-util.h"    // u64, u32 and perf_mmap_data_t
 
 /******************************************************************************
  * perfmon
@@ -227,7 +228,7 @@ show_info(char *event )
 // return 0 or positive if the event exists, -1 otherwise
 // if the event exist, code and type are the code and type of the event
 int 
-pfmu_getEventType(const char *eventname, unsigned int *code, unsigned int *type)
+pfmu_getEventType(const char *eventname, u64 *code, u64 *type)
 {
   pfm_perf_encode_arg_t arg;
   char *fqstr = NULL;
@@ -258,7 +259,7 @@ pfmu_getEventType(const char *eventname, unsigned int *code, unsigned int *type)
 int
 pfmu_isSupported(const char *eventname)
 {
-  unsigned int eventcode, eventtype;
+  u64 eventcode, eventtype;
   return pfmu_getEventType(eventname, &eventcode, &eventtype);
 }
